UI Revisions
Ordering claims by start datetime, aligning tables, cleaning css.
