+++ src/WebUi/Areas/Underwriter/Models/ClaimsListViewModel.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using EDeviceClaims.Core;
 using EDeviceClaims.Domain.Models;
 
@@ -14,28 +13,5 @@ namespace EDeviceClaims.WebUi.Areas.Underwriter.Models
                 Add(new UnderwriterClaimViewModel(claim));
             }
         }
-    }
-
-    public class UnderwriterClaimViewModel
-    {
-        public string Status { get; set; }
-
-        public string Start { get; set; }
-
-        public string Name { get; set; }
-
-        public Guid PolicyId { get; set; }
-
-        public Guid Id { get; set; }
-
-        public UnderwriterClaimViewModel(ClaimDomainModel claim)
-        {
-            Id = claim.Id;
-            PolicyId = claim.Policy.Id;
-            Name = claim.Policy.DeviceName;
-            Start = $"{claim.WhenStarted.ToShortDateString()} {claim.WhenStarted.ToShortTimeString()}";
-            Status = claim.Status.ToString();
-        }
-        
     }
 }
