+++ src/WebUi/Areas/Underwriter/Views/Claims/Index.cshtml
@@ -1,27 +1,72 @@
 ﻿@model EDeviceClaims.WebUi.Areas.Underwriter.Models.ClaimsListViewModel
 
-<h1>All Open Claims</h1>
+<h1 align="center">Current Open Claims</h1>
 
-<table>
-    <tr>
-        <th>Claim Id</th>
-        <th>Policy Id</th>
-        <th>Device</th>
-        <th>Customer</th>
-        <th>Claim Date</th>
-        <th>Status</th>
-    </tr>
+
+
     
-    @foreach (var claim in Model)
-    {
+    <table>
         <tr>
-            <td>@claim.Id</td>
-            <td>@claim.PolicyId</td>
-            <td>@claim.Name</td>
-            <td>Customer</td>
-            <td>@claim.Start</td>
-            <td>@claim.Status</td>
+            <th>Claim Id</th>
+            <th>Claim Date</th>
+            <th>Actions</th>
         </tr>
+
+        @foreach (var claim in Model)
+        {
+            <tr>
+                <td>@claim.Id</td>
+                <td>@claim.Start</td>
+                <td>@Html.ActionLink("View Details", "Details", "Claims", routeValues: new {claimId = claim.Id}, htmlAttributes: new {Class = "btn btn-primary btn-lg"})</td>
+            </tr>
+        }
+
+    </table>
+
+
+<style>
+    
+    #tableWrapper {
+        text-align: center;
+        padding-top: 25px;
+        left: 50%;
+        margin-right: -50%;
+        transform: translate(-50%, -50%);
     }
 
-</table>
+     table#claimListTable {
+         margin: auto;
+         width: 75%;
+         background-color: #f1f1c1;
+         text-align: center;
+     }
+
+    table#claimListTable tr:nth-child(even) {
+        background-color: #eee;
+    }
+
+    table#claimListTable tr:nth-child(odd) {
+        background-color: #fff;
+    }
+
+    #claimListTable th {
+        color: black;
+        background-color: gray;
+        text-align: center;
+    }
+
+    table, th, td {
+        border: 1px solid black;
+        border-collapse: collapse;
+    }
+
+    th, td {
+        padding: 15px;
+        text-align: center;
+    }
+    table {
+        margin-left: auto;
+        margin-right: auto;
+        width: 75%;
+    }
+</style>
