+++ src/WebUi/Views/Devices/Index.cshtml
@@ -3,9 +3,7 @@
 <h1 align="center">Your Insured Devices</h1>
 
 
-<div>
-
-    <table class="table" id="deviceListTable">
+<table class="table" id="deviceListTable">
         <tr>
             <th>@Html.Label("Device")</th>
             <th>@Html.Label("Serial Number")</th>
@@ -25,8 +23,6 @@
         }
     </table>
 
-</div>    
-
 <style>
     
     table#deviceListTable { margin: auto; width: 75%; background-color: #f1f1c1; text-align: center; }
@@ -36,10 +32,6 @@
     table#deviceListTable tr:nth-child(odd) { background-color: #fff; }
 
     table#deviceListTable th { color: black; background-color: gray; text-align: center; }
-
-    table, th, td { border: 1px solid black; }
-
-    table, th, td { border: 1px solid black; border-collapse: collapse; }
 
     table, th, td { border: 1px solid black; border-collapse: collapse; }
