+++ src/WebUi/WebUi.csproj
@@ -170,6 +170,7 @@
     <Compile Include="Areas\Underwriter\Controllers\ClaimsController.cs" />
     <Compile Include="Areas\Underwriter\Controllers\UnderwriterAppController.cs" />
     <Compile Include="Areas\Underwriter\Models\ClaimsListViewModel.cs" />
+    <Compile Include="Areas\Underwriter\Models\UnderwriterClaimViewModel.cs" />
     <Compile Include="Areas\Underwriter\UnderwriterAreaRegistration.cs" />
     <Compile Include="Controllers\AccountController.cs" />
     <Compile Include="Controllers\AppController.cs" />
@@ -293,6 +294,8 @@
     <Content Include="Areas\Underwriter\Views\web.config" />
     <Content Include="Areas\Underwriter\Views\Claims\Index.cshtml" />
     <Content Include="Areas\Underwriter\Views\_ViewStart.cshtml" />
+    <Content Include="Areas\Underwriter\Views\Claims\Details.cshtml" />
+    <Content Include="Areas\Underwriter\Views\Claims\_claimDetailsTable.cshtml" />
     <None Include="Scripts\jquery-1.10.2.intellisense.js" />
     <Content Include="Scripts\device_details.js" />
     <Content Include="Scripts\indexJavaScript.js" />
